Make RPCModules generic to allow beacon modules
RPCModules need to access the chain regulary
so it makes sense for them to depend and provide
access to the chain. However, the RPCServer
doesn't care about the concrete chain type
and we can get away with making it generic
since Eth1 and Beacon chain don't share a
common ancestor.