Don't implicitly create process ProcessPoolExecutor
Previously, every process (e.g BaseIsolatedPlugin)
implicitly created a ProcessPoolExecutor with n
processes where n = HOST CPU COUNT - 1

This changes the model so that the `BaseService` does
not depend on a globally created executor anymore.

The executor lifecycle is handled by the `networking`
process which is currently the only place that uses
it anyway.