Refactor how PeerPool listens for external EventBus requests
With this change, external requests to the PeerPool
can be defined by subclassing `BasePeerPoolEventBusAPI`
and setting that as `event_bus_api_class` on the specific
PeerPool class. This is symmetric with the handling of
the `peer_factory_class`.

This avoids polluting the PeerPool with lots of code
to handle specific requests and also opens up
extensibility in regard to the various different forms
of requests that the different concrete `PeerPool`
instances need to learn how to handle.

This is related to #213