Refactor how PeerPool listens for external EventBus requests
With this change, external requests to the PeerPool
are handled by a BasePeerPoolEventBusRequestHandler
which can be subclassed to handle non-base events.

This also moves all p2p.events that are not intrinsically
needed by the PeerPool to
trinity.protocol.common.events.

This avoids polluting the PeerPool with lots of code
to handle specific requests and also opens up
extensibility in regard to the various different forms
of requests that the different concrete `PeerPool`
instances need to learn how to handle.

This is related to #213