p2p/{peer,peer_pool}: debug-print remote client version strings.
This gives an additional insight into remotes, and is a first step
in debugging possible remote-specific failures.

The string-to-print part that will be retained in memory is limited
in size; and is minimally "sanitised". First, tabs/newlines and other
"whitespace" is stripped (always); if the string is still unprintable at
this point, then illegal characters are "replaced" with question marks:

https://docs.python.org/3/howto/unicode.html#the-string-type

(search `replace`).