Expose peer cmds for new block hashes and transactions on event bus
These commands need to be exposed on the event bus to
become available to isolated plugins and hence allow
things like the transaction pool to be moved into an
isolated plugin.

Notice that exposing these events does *not* cause extra
traffic on the event bus per se. They are only broadcasted
to those endpoints that are actually listening for these
events.