Fix flaky test
We currently wait for the json-rpc Endpoint to connect
to other endpoints before we do certain assertions.
But sometimes we might end up in a situation where
al other endpoints have already connected to this
endpoint and so it does not actively connect to
others.

It is safer to simply wait for the announcement
of the JSON-RPC endpoint instead.

Fixes #1362