eth1_monitor: fix endianness when storing in DB.
With the `web3.py` v4->v5 change, the `log.amount` was now passed as
`bytes`, so there's a chance there's something wrong with the ABI.
(When `web3` v4 was being used, this type conversion was not needed.)

Plain conversion to `int` results is test errors: wrong amount, and
subsequently wrong data (local DB deposit data corrupted). This is
because the data obtained from the log / event object, as emitted
by the contract, is of different endianness.

To fix this and arising typing issues, instead use a "dead code"
function from `DepositLog` and a double-pass to first obtain/process
the logs (raw data), and then populate objects (with typing).