Fix discv4 to not ping same node more than once at a given time
DiscoveryService was wrongly using the CallbackManager that is supposed
to prevent us from pinging a node which we already pinged and for which
we're currently waiting for a reply. In some places it was passing the
expected reply token as the key to CallbackManager and in others it was
passing the remote itself. Now we always use the remote, and
CallbackManager's method signatures have also been changed to expect a
NodeAPi as the key instead of a Hashable.