Remove "remove" method from ENR DB API
The problem is that the only parameter to "remove" is the node id, which
is not a unique identifier for an ENR (we'd need a sequence number as
well). Currently, the memory db only stores a single ENR per node, so
this isn't a problem. But the future file enr db will also keep
historical records. Therefore, I've removed the method from the API (but
kept the implementation).