PeerBackends now take a function to filter candidates
Instead of simply skipping peers we are already connected to,
PeerBackends now take a `should_skip(NodeAPI) -> bool` function.

This way we avoid a situation in which the PeerPool would repeatedly
receive connection candidates that are known to be incompatible (e.g.
peers that have been blacklisted because of a proto/chain mismatch or
whose forkid is not compatible with ours).