Every component's process is now its own process group leader
IOW, for every process we start, we create a new process group.
This way a Ctrl-C in the terminal won't cause a SIGINT to be sent
to each of our processes, instead a single one will be sent to the
parent process, which will then send SIGINTs to every child (via
asyncio-run-in-process' open_in_process()).