Make sure _handle_shutdown_request() doesn't exit early
The ComponentManager runs _handle_shutdown_request() as a daemon, but it
would exit as soon as it received the first shutdown request, causing
asyncio-service to raise a DaemonTaskExit. Now it will loop for as long
as the manager is running and ignore (albeit with a log warning) any
requests but the first one.

Also change callsites that emit a ShutdownRequest() to use
broadcast_nowait(), to avoid potential hangs