Fix logging on the main process
The root logger of the main process was configured with the level
used for the stderr handler, so any logs from the main process with
severity lower than that would be filtered before reaching the file
handler. Now we use the minimum level on the root logger of the
main process, like we do for the root logger on child processes.

Closes: #1668