Clip the length of really long DEBUG2 logs
This was often causing test_logger_configuration() to fail because
DEBUG2 printout of the command was huge (the "Discarding..." log and the
"sync received new headers" logs were called particularly often).

The printout was so big that the async stream reader would fail. More
details in the comment near the new maximum-length constant.