Bugfix: beam sync hung when only peer is penalized
A call to QueeningQueue.get_queen_peer() is meant to hang when there
is no queen. This fixed bug meant that after a queen was penalized, and
then reinserted at the end of a penalty, the get_queen_peer() call would
hang until a second peer was inserted and retrieved by
pop_fastest_peasant(). That means it hangs forever if you are only
connected to a single peer and it gets penalized.