Increase hypothesis deadline for CI variance
CI took 240ms on one run, then 1ms on the next, with the same input.

________________________ test_random_pruning[3-1-False] ________________________

---------------------------------- Hypothesis ----------------------------------

Falsifying example: test_random_pruning(
    batch_size=3,
    ignore_duplicates=False,
    prune_depth=3,
    recomplete_idx=1,
    task_series=[4, 4, 3, 3, 2, 0, 5, 1, 1, 2, 5, 5, 5],
)

Unreliable test timings! On an initial run, this test took 264.00ms, which exceeded the deadline of 200.00ms, but on a subsequent run it took 1.08 ms, which did not. If you expect this sort of variability in your test timings, consider turning deadlines off for this test by setting deadline=None.