Ensure the Syncer component exits if the Node service does so
The Syncer used to run the Node service in the background and wait
on the sync task (which runs forever) before calling
node_manager.wait(). As a result, if the Node crashed the sync task
would continue running but do nothing because the PeerPool (part of
the Node's functionalities) would never have any peers.