Ignore delay when inserting a queen-able peasant
If we are blocked waiting for a queen, we want to resolve the situation
as quickly as possible. (or if we have a candidate who is a faster
queen)

So if we are inserting a peasant, but it will be promoted to queen, then
insert it immediately, even if the insert was called with a delay.