Narrower block backfill range, to prefer recency
If we have a single giant block gap, then just picking the latest gap
isn't enough.

We could try to download all the blocks in reverse, but the syncers are
not built to go in reverse order. Instead, we get a narrower block
range than the full gap, preferring the blocks closest to the tip.