Backfill more than one storage node per round
The storage iterator would exit when a node was found and on the next
attempt to iterate, it would exit. So we were only ever getting a single
missing storage node per round. Now, we group them up as intended. This
should make backfill focus more tightly on the index key. I'm hoping to
see walked ppm increase a bit faster in the normal case.

This also really helps during the backfill tests, which tend to get
stuck on a single large storage account at the end. It only has one peer
to ask, and it then delays the peer by 250ms after each request. So it
was kind of a miracle that backfill was ever finishing during the test.