Discovery: msg handlers now await for replies to be sent
Our msg handlers already run as background tasks, so it made no sense for
them to run the coroutine that sends the reply over the socket in the
background as well.

Also fixes request_enr() and _lookup() so that they only start waiting
for replies after the ENR/FIND_NODE msgs have been sent.