Fix flaky P2P server tests
A couple tests could sometimes attempt to connect to a server
before it was actually listening, and since the PeerPool
suppresses the UnreachableError (as it is expected), the tests would
then fail with a TimeoutError, Now the test fixture will only yield
once the server is listening and the PeerPool used in tests does
not suppress any errors when connecting.

Also better logs on PeerPool to help debugging and better asyncio
task names for connection behaviors.

Closes: #1504