Stabilize a flaky log-over-IPC test
Not convinced this really fixed it, but I think it gives extra time for
the logs to make it to the handler. Ideally, the server shouldn't shut
down until all the logs are passed on, but I couldn't confirm that with
a quick scan.

test_queued_logging failed once at:

        with queue_listener.run(ipc_path):
            assert len(handler.logs) == 0
            proc.start()
            proc.join()
>           assert len(handler.logs) == 3
E           assert 0 == 3
E            +  where 0 = len([])
E            +    where [] = <HandlerForTest (DEBUG)>.logs