Stabilize another token bucket test
test_token_bucket_refills_itself was failing because the zero measure
was only taking 0.00003 seconds, and the actual loop time was 0.00017.

These numbers are so small that it's pretty meaningless. We might as
well add a hard-coded minimum epsilon that it's okay to be shorter than.
Just in case it's not, we make sure that it's shorter than the slowest
of 5 "empty" loops.